Move all plugins into explicit latest directory
Since we're only storing one version of plugins, it no longer makes
sense to auto-generate the "latest" version of a plugin, as this results
in e.g. two versions for each plugin

- redhat/vscode-java11/1.7.0
- redhat/vscode-java11/latest

(which are identical). In the case that a user selects version 1.7.0
explicitly, the only difference is that their workspace breaks if the
java plugin is updated.

This commit disables auto-generating latest meta.yamls by:

* Updating script generate_latest_metas.sh to not require a latest.txt to
exist for each plugin
* Moving each plugin that is currently set as "latest" (according to
latest.txt) to the ./latest directory for that plugin
* Updating version for each meta.yaml in a latest directory to be "latest"
* Updating gitignore to no longer exclude latest metas for each plugin

Note: This commit leaves the mechanism for automatically generating
latest plugin meta.yamls in place; if a directory contains a
"latest.txt", the previous behavior will occur.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>